Move `ConnectionProxy` into `ServiceEndpoint`
