+++ android/src/main/kotlin/net/mullvad/mullvadvpn/service/MullvadVpnService.kt
@@ -50,6 +50,9 @@ class MullvadVpnService : TalpidVpnService() {
     private val binder = LocalBinder()
     private val serviceNotifier = EventNotifier<ServiceInstance?>(null)
 
+    private val connectionProxy
+        get() = endpoint.connectionProxy
+
     private var state = State.Running
 
     private var setUpDaemonJob: Job? = null
@@ -76,16 +79,13 @@ class MullvadVpnService : TalpidVpnService() {
     private lateinit var keyguardManager: KeyguardManager
     private lateinit var notificationManager: ForegroundNotificationManager
     private lateinit var tunnelStateUpdater: TunnelStateUpdater
-    private lateinit var vpnPermission: VpnPermission
 
     private var pendingAction by observable<PendingAction?>(null) { _, _, _ ->
-        val connectionProxy = instance?.connectionProxy
-
         // The service instance awaits the split tunneling initialization, which also starts the
         // endpoint. So if the instance is not null, the endpoint has certainly been initialized.
-        if (connectionProxy != null) {
+        if (instance != null) {
             endpoint.settingsListener.settings?.let { settings ->
-                handlePendingAction(connectionProxy, settings)
+                handlePendingAction(settings)
             }
         }
     }
@@ -110,7 +110,8 @@ class MullvadVpnService : TalpidVpnService() {
             Looper.getMainLooper(),
             daemonInstance.intermittentDaemon,
             connectivityListener,
-            SplitTunnelingPersistence(this)
+            SplitTunnelingPersistence(this),
+            VpnPermission(this)
         )
 
         notificationManager =
@@ -126,8 +127,6 @@ class MullvadVpnService : TalpidVpnService() {
 
             start()
         }
-
-        vpnPermission = VpnPermission(this)
     }
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
@@ -234,7 +233,6 @@ class MullvadVpnService : TalpidVpnService() {
     }
 
     private suspend fun setUpInstance(daemon: MullvadDaemon, settings: Settings) {
-        val connectionProxy = ConnectionProxy(vpnPermission, daemonInstance.intermittentDaemon)
         val customDns = CustomDns(daemon, endpoint.settingsListener)
 
         endpoint.splitTunneling.onChange.subscribe(this@MullvadVpnService) { excludedApps ->
@@ -243,7 +241,7 @@ class MullvadVpnService : TalpidVpnService() {
             connectionProxy.reconnect()
         }
 
-        handlePendingAction(connectionProxy, settings)
+        handlePendingAction(settings)
 
         endpoint.locationInfoCache.stateEvents = connectionProxy.onStateChange
 
@@ -280,7 +278,7 @@ class MullvadVpnService : TalpidVpnService() {
         }
     }
 
-    private fun handlePendingAction(connectionProxy: ConnectionProxy, settings: Settings) {
+    private fun handlePendingAction(settings: Settings) {
         when (pendingAction) {
             PendingAction.Connect -> {
                 if (settings.accountToken != null) {
